refactor watchDog into new class
